requirements.txt
# Core dependencies
flask>=2.0.0
ultralytics>=8.0.0
pillow>=9.0.0
numpy>=1.20.0

# File handling and security
werkzeug>=2.0.0

# Optional: For better logging and monitoring
python-dotenv>=0.19.0

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# pytest-flask>=1.2.0
# black>=22.0.0
# flake8>=4.0.0
